DAT-804 refactor search module
