+++ app/src/app/main/main-routing.module.ts
@@ -49,7 +49,7 @@ const routes: Routes = [
         },
         canActivate: [AuthGuard],
         canActivateChild: [AuthGuard],
-        loadChildren: () => import('../search/search.module').then(m => m.SearchModule)
+        loadChildren: () => import('../modules/search/search.module').then(m => m.SearchModule)
     },
     {
         path: 'framework',
