+++ app/src/app/modules/search-bar/components/search-bar-tag-icon/search-bar-tag-icon.component.html
@@ -1,0 +1,8 @@
+<ng-container [ngSwitch]="tag.searchForm">
+    <i class="fas fa-font" *ngSwitchCase="'text'"></i>
+    <i class="fas fa-code" *ngSwitchCase="'regex'"></i>
+    <i class="fas fa-qrcode" *ngSwitchCase="'publicID'"></i>
+    <i class="fas fa-object-group" *ngSwitchCase="'type'"></i>
+    <i class="fas fa-folder-open" *ngSwitchCase="'category'"></i>
+    <i class="fas fa-font" *ngSwitchDefault></i>
+</ng-container>
