+++ app/src/app/modules/search-bar/components/search-bar-tag/search-bar-tag.component.html
@@ -1,0 +1,12 @@
+<div class="btn-group">
+    <button type="button" class="btn btn-primary btn-sm search-bar-tag-icon dropdown-toggle-split disabled"
+            data-toggle="dropdown">
+        <cmdb-search-bar-tag-icon [tag]="tag"></cmdb-search-bar-tag-icon>
+    </button>
+    <button type="button" class="btn btn-primary btn-sm" data-toggle="dropdown" #dropdownTrigger
+            data-reference="parent">
+        <ng-content></ng-content>
+    </button>
+    <button (click)="emitDelete()" type="button" class="btn btn-primary btn-sm dropdown-toggle dropdown-toggle-split">
+    </button>
+</div>
