+++ app/src/app/modules/search-bar/components/search-bar-tag/search-bar-tag.component.scss
@@ -1,0 +1,32 @@
+$dropdown-form-width: 250px;
+
+form.dropdown-menu {
+  width: $dropdown-form-width;
+}
+
+button {
+  position: relative;
+  display: inline-block;
+  border-radius: .15rem;
+  padding: .3rem .6rem .3rem .6rem;
+  font-size: 75%;
+  font-weight: 700;
+  color: white;
+  line-height: 1;
+
+  &.search-bar-tag-icon {
+    display: inline-block;
+    border: 0;
+    padding-right: 0.375rem;
+    padding-left: 0.375rem;
+  }
+
+  &.dropdown-toggle {
+
+    &::after {
+      content: "\00d7";
+      border: 0;
+      font-size: 120%;
+    }
+  }
+}
