+++ app/src/app/modules/search/components/search-result-match/search-result-match.component.html
@@ -1,0 +1,3 @@
+<ng-container *ngIf="matchedField">
+    <i>{{matchedField.label}}</i>: <cmdb-render-element [mode]="MODE" [data]="matchedField"></cmdb-render-element>
+</ng-container>
