+++ app/src/app/modules/search/components/search-result/search-result.component.scss
