--- app/src/app/search/search-bar/search-bar-tag-settings-form/search-bar-tag-settings-form.component.html
@@ -1,3 +1,0 @@
-<ng-container [formGroup]="settingsControl">
-
-</ng-container>
