--- app/src/app/search/search-bar/search-bar-tag/search-bar-tag-icon.component.html
@@ -1,8 +1,0 @@
-<ng-container [ngSwitch]="tag.searchForm">
-  <i class="fas fa-font" *ngSwitchCase="'text'"></i>
-  <i class="fas fa-code" *ngSwitchCase="'regex'"></i>
-  <i class="fas fa-qrcode" *ngSwitchCase="'publicID'"></i>
-  <i class="fas fa-object-group" *ngSwitchCase="'type'"></i>
-  <i class="fas fa-folder-open" *ngSwitchCase="'category'"></i>
-  <i class="fas fa-font" *ngSwitchDefault></i>
-</ng-container>
