--- app/src/app/search/search-bar/search-bar-tag/search-bar-tag.component.scss
@@ -1,48 +1,0 @@
-$dropdown-form-width: 250px;
-
-form.dropdown-menu {
-  width: $dropdown-form-width;
-}
-
-button {
-  position: relative;
-  display: inline-block;
-  border-radius: .15rem;
-  padding: .3rem .6rem .3rem .6rem;
-  font-size: 75%;
-  font-weight: 700;
-  color: white;
-  line-height: 1;
-
-  &.search-bar-tag-icon {
-    display: inline-block;
-    border: 0;
-    padding-right: 0.375rem;
-    padding-left: 0.375rem;
-  }
-
-  &.dropdown-toggle {
-
-    &::after {
-      content: "\00d7";
-      border: 0;
-      font-size: 120%;
-    }
-  }
-
-  /*button {
-    background-color: transparent;
-    border: 0;
-    opacity: .9;
-    color: inherit;
-    font-size: 1.5em;
-    line-height: 1;
-    position: absolute;
-
-    &.delete {
-      top: 0;
-      right: 0;
-      padding: 0.1rem 0.4rem;
-    }
-  }*/
-}
