--- app/src/app/search/search-result/search-result-bar/search-result-bar.component.html
@@ -1,53 +1,0 @@
-
-<div class="container-fluid">
-  <div class="row object-view-navbar">
-    <div class="col-2">
-      <div class="d-flex justify-content-start">
-        <div class="icon"><i class="far fa-list-alt"></i></div>
-        <div class="name">
-          <strong class="text-uppercase">Results</strong><br>
-          <span> Total {{searchResultList?.total_results}}</span>
-        </div>
-      </div>
-    </div>
-    <div class="col-8 border-left d-flex justify-content-start">
-      <div class="d-flex">
-        <div class="icon"><i class="fas fa-filter mr-1"></i></div>
-        <div class="name"><strong class="text-uppercase">Filter</strong>
-          <ng-select [items]="filterResultList"
-                     bindLabel="searchLabel"
-                     multiple="true"
-                     placeholder="Select filter"
-                     [(ngModel)]="preSelectedFilterList"
-                     (clear)="rollbackQueryParametersIfNeeded()"
-                     (add)="addFilterItem($event)"
-                     (remove)="delFilterItem($event.value)">
-            <ng-template ng-label-tmp let-item="item" let-clear="clear">
-              <span class="ng-value-icon right">{{item.total}}</span>
-              <span class="ng-value-label">{{item.searchLabel}}</span>
-              <span class="ng-value-icon left" (click)="clear(item)">×</span>
-            </ng-template>
-            <ng-template ng-option-tmp let-item="item" let-index="index">
-              <div class="d-flex justify-content-between align-items-center">
-                <span class="text-wrap" title="{{item.searchLabel}}">{{item.searchLabel}}</span>
-                <span class="badge badge-primary badge-pill">{{item.total}}</span>
-              </div>
-            </ng-template>
-          </ng-select>
-        </div>
-      </div>
-    </div>
-    <div class="col-2 border-left">
-      <div class="d-flex justify-content-center">
-        <div class="icon"><i class="fas fa-share-alt"></i></div>
-        <div class="name">
-          <strong class="text-uppercase">Share URL</strong><br>
-          <button type="button" title="Copy filter URL"
-                  class="btn btn-outline-secondary btn-sm btn-block" (click)="copyToClipboard()">
-            <i class="far fa-copy"></i>
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
