--- app/src/app/search/search-result/search-result-match.component.html
@@ -1,3 +1,0 @@
-<ng-container *ngIf="matchedField">
-  <i>{{matchedField.label}}</i>: <cmdb-render-element [mode]="MODE" [data]="matchedField"></cmdb-render-element>
-</ng-container>
