--- app/src/app/search/search-result/search-result-preview.component.html
@@ -1,14 +1,0 @@
-<div class="modal-header">
-  <h4 class="modal-title">Preview</h4>
-  <button type="button" class="close" aria-label="Close" (click)="activeModal.dismiss()">
-    <span aria-hidden="true">&times;</span>
-  </button>
-</div>
-<div class="modal-body">
-  <div class="renderForm">
-    <cmdb-render [renderResult]="renderResult" [mode]="mode" [renderForm]="formGroupDummy"></cmdb-render>
-  </div>
-</div>
-<div class="modal-footer">
-  <button type="button" class="btn btn-outline-dark" (click)="activeModal.close()">Close</button>
-</div>
